Move runMain into TestCase
